Adequando ao padrão mv - alocando classes em seus respectivos diretorios
