+++ src/main/java/br/upe/ppsw/jabberpoint/JabberPointApplication.java
@@ -27,11 +27,12 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
-import br.upe.ppsw.jabberpoint.apresentacao.DataAccessor;
-import br.upe.ppsw.jabberpoint.apresentacao.Presentation;
-import br.upe.ppsw.jabberpoint.apresentacao.SlideViewerFrame;
-import br.upe.ppsw.jabberpoint.apresentacao.Style;
-import br.upe.ppsw.jabberpoint.apresentacao.XMLAccessor;
+
+import br.upe.ppsw.jabberpoint.apresentacao.model.DataAccessor;
+import br.upe.ppsw.jabberpoint.apresentacao.model.Presentation;
+import br.upe.ppsw.jabberpoint.apresentacao.model.XMLAccessor;
+import br.upe.ppsw.jabberpoint.apresentacao.view.SlideViewerFrame;
+import br.upe.ppsw.jabberpoint.apresentacao.view.Style;
 
 @SpringBootApplication
 public class JabberPointApplication implements CommandLineRunner {
