+++ src/main/java/br/upe/ppsw/jabberpoint/apresentacao/controller/KeyController.java
@@ -18,10 +18,11 @@
  * 
  * @author Ian F. Darwin, hbarreiros
  */
-package br.upe.ppsw.jabberpoint.apresentacao;
+package br.upe.ppsw.jabberpoint.apresentacao.controller;
 
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
+import br.upe.ppsw.jabberpoint.apresentacao.model.Presentation;;
 
 public class KeyController extends KeyAdapter {
