+++ src/main/java/br/upe/ppsw/jabberpoint/apresentacao/controller/MenuController.java
@@ -18,9 +18,10 @@
  * 
  * @author Ian F. Darwin, hbarreiros
  */
-package br.upe.ppsw.jabberpoint.apresentacao;
+package br.upe.ppsw.jabberpoint.apresentacao.controller;
 
 import java.awt.Frame;
+
 import java.awt.Menu;
 import java.awt.MenuBar;
 import java.awt.MenuItem;
@@ -30,6 +31,10 @@ import java.awt.event.ActionListener;
 import java.io.IOException;
 import javax.swing.JOptionPane;
 import org.springframework.util.ResourceUtils;
+
+import br.upe.ppsw.jabberpoint.apresentacao.model.Presentation;
+import br.upe.ppsw.jabberpoint.apresentacao.view.AboutBox;
+import br.upe.ppsw.jabberpoint.apresentacao.model.*;
 
 public class MenuController extends MenuBar {
