+++ src/main/java/br/upe/ppsw/jabberpoint/apresentacao/model/DataAccessor.java
@@ -18,7 +18,7 @@
  * 
  * @author Ian F. Darwin, hbarreiros
  */
-package br.upe.ppsw.jabberpoint.apresentacao;
+package br.upe.ppsw.jabberpoint.apresentacao.model;
 
 import java.io.IOException;
