+++ src/main/java/br/upe/ppsw/jabberpoint/apresentacao/model/DemoData.java
@@ -18,7 +18,7 @@
  * 
  * @author Ian F. Darwin, hbarreiros
  */
-package br.upe.ppsw.jabberpoint.apresentacao;
+package br.upe.ppsw.jabberpoint.apresentacao.model;
 
 import java.io.FileNotFoundException;
 import org.springframework.util.ResourceUtils;
