+++ src/main/java/br/upe/ppsw/jabberpoint/apresentacao/model/TextItem.java
@@ -18,7 +18,7 @@
  * 
  * @author Ian F. Darwin, hbarreiros
  */
-package br.upe.ppsw.jabberpoint.apresentacao;
+package br.upe.ppsw.jabberpoint.apresentacao.model;
 
 import java.awt.Graphics;
 import java.awt.Graphics2D;
@@ -34,6 +34,8 @@ import java.text.AttributedString;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+
+import br.upe.ppsw.jabberpoint.apresentacao.view.Style;
 
 public class TextItem extends SlideItem {
